Update tests to detect local address, remove OS X specific code.
